requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
openai==1.58.1
python-dotenv==1.0.1
pydantic-settings==2.7.1
python-multipart==0.0.20
PyMuPDF==1.25.3
python-docx==1.1.2
openpyxl==3.1.5
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
